chore: 프로필이미지 크롭 & 반려동물 등록 모달 커스텀 훅/컴포넌트 분할
