+++ src/components/modals/ProfileImageCropModal.tsx
@@ -1,12 +1,11 @@
-import { useEffect, useRef, useState } from "react";
 import Cropper from "react-cropper";
-import type { ReactCropperElement } from "react-cropper";
 import "@/styles/ImageCropModal.css";
 import ModalBackground from "./common/ModalBackground";
 import DefaultButtonCancel from "../common/DefaultButtonCancel";
 import { useModalEscapeKey } from "@/hooks/useModalEscapeKey";
 import DefaultButtonConfirm from "../common/DefaultButtonConfirm";
 import { changeProfileImage } from "@/lib/myInfoUtils";
+import { useProfileImageCropModal } from "@/hooks/useProfileImageCropModal";
 
 interface Props {
   onClose: () => void;
@@ -17,78 +16,17 @@ export default function ProfileImageCropModal({
   onClose,
   currentImage,
 }: Props) {
-  const cropperRef = useRef<ReactCropperElement>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const [imageURL, setImageURL] = useState<string>(""); // 원본 미리보기용
-  const [croppedFile, setCroppedFile] = useState<File | null>(null); // ⬅️ 크롭 결과
-  const [previewURL, setPreviewURL] = useState<string>(""); // ⬅️ 크롭 결과 미리보기용
-
   useModalEscapeKey(onClose);
-
-  useEffect(() => {
-    if (currentImage) setImageURL(currentImage);
-    return () => {
-      if (imageURL.startsWith("blob:")) URL.revokeObjectURL(imageURL);
-      if (previewURL.startsWith("blob:")) URL.revokeObjectURL(previewURL);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  // 크롭 결과 파일이 바뀔 때마다 미리보기 URL 갱신
-  useEffect(() => {
-    if (previewURL.startsWith("blob:")) URL.revokeObjectURL(previewURL);
-    if (croppedFile) {
-      const url = URL.createObjectURL(croppedFile);
-      setPreviewURL(url);
-    } else {
-      setPreviewURL("");
-    }
-  }, [croppedFile]);
-
-  const handlePick = () => inputRef.current?.click();
-
-  const onFileChange: React.ChangeEventHandler<HTMLInputElement> = (e) => {
-    const f = e.target.files?.[0];
-    if (!f) return;
-
-    if (imageURL.startsWith("blob:")) URL.revokeObjectURL(imageURL);
-    const url = URL.createObjectURL(f);
-    setImageURL(url);
-    setCroppedFile(null); // 새 파일 선택 시 이전 크롭 결과 초기화
-    e.currentTarget.value = ""; // 같은 파일 재선택 가능
-  };
-
-  const handleCrop = () => {
-    const cropper = cropperRef.current?.cropper;
-    if (!cropper || !imageURL) return;
-
-    const canvas = cropper.getCroppedCanvas({
-      width: 300,
-      height: 300,
-      imageSmoothingQuality: "high",
-    });
-
-    if (!canvas) {
-      alert("이미지를 크롭할 수 없습니다. 다른 이미지를 시도해 주세요.");
-      return;
-    }
-
-    // ⬇️ dataURL 대신 Blob/File로 받기
-    canvas.toBlob(
-      (blob) => {
-        if (!blob) {
-          alert("이미지 변환에 실패했습니다.");
-          return;
-        }
-        const file = new File([blob], "avatar.jpg", { type: "image/jpeg" });
-        setCroppedFile(file);
-      },
-      "image/jpeg",
-      0.9
-    );
-  };
-
+  const {
+    inputRef,
+    onFileChange,
+    imageURL,
+    handlePick,
+    cropperRef,
+    previewURL,
+    croppedFile,
+    handleCrop,
+  } = useProfileImageCropModal({ currentImage });
   return (
     <ModalBackground onClose={onClose}>
       <div
@@ -179,8 +117,7 @@ export default function ProfileImageCropModal({
             textSize={14}
             onClick={() => {
               if (!croppedFile) return;
-              // changeProfileImage가 string을 받던 함수였다면, 여기서 FormData로 업로드하도록 그 함수 시그니처를 바꿔줘.
-              // 예: changeProfileImage(croppedFile, user?.name!)
+
               changeProfileImage(croppedFile);
             }}
             isActive={!!croppedFile}
